Fix: Resolve compilation errors in SupabaseDatabaseService and ActivityResultHandler
Refactored the `SupabaseQuery` class to correctly build queries using the `select` builder pattern, resolving unresolved reference errors for `order`, `limit`, and `lt`.

Fixed a type mismatch in `ActivityResultHandler` by casting an `Int` to a `Double` when calling `_startUploadForItem`.

+++ app/src/main/java/com/synapse/social/studioasinc/util/ActivityResultHandler.kt
@@ -80,7 +80,7 @@ class ActivityResultHandler(private val activity: ChatActivity) {
                     // Start upload for each item
                     for (i in resolvedFilePaths.indices) {
                         try {
-                            activity._startUploadForItem(startingPosition + i)
+                            activity._startUploadForItem((startingPosition + i).toDouble())
                         } catch (e: Exception) {
                             Log.e("ChatActivity", "Error starting upload for item " + i + ": " + e.message)
                         }
